Version 0.002
the "SVN commit" and Commit methods now open up a window inline, manually fetching the repository status and generating a little file exactly like the one svn generates, then a new 'Complete' command was added that wraps a small sh script around the commit information you provided, which is then given to svn commit as the editor, and then removed.  The window is then closed, and its all done quietly.  I also added 'Svn log' and 'Log' commands that open up the log in a new window and move your cursor to the top :)  This thing is almost cool now